+++ src/main/java/space/space_spring/config/WebConfig.java
@@ -1,0 +1,22 @@
+package space.space_spring.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import space.space_spring.interceptor.JwtAuthInterceptor;
+import space.space_spring.jwt.JwtProvider;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig implements WebMvcConfigurer {
+
+    private final JwtAuthInterceptor jwtAuthInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(jwtAuthInterceptor)
+                .order(1)
+                .addPathPatterns("/space/**", "/test/**");          // interceptor 적용되야하는 url enum으로 만들어서 여기에 달면 될듯
+    }
+}
