+++ src/main/java/space/space_spring/service/UserService.java
@@ -3,11 +3,10 @@ package space.space_spring.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import space.space_spring.JwtProvider;
+import space.space_spring.jwt.JwtProvider;
 import space.space_spring.dao.UserDao;
 import space.space_spring.domain.User;
 import space.space_spring.dto.PostUserLoginRequest;
-import space.space_spring.dto.PostUserLoginResponse;
 import space.space_spring.dto.PostUserSignupRequest;
 import space.space_spring.dto.PostUserSignupResponse;
 import space.space_spring.exception.UserException;
